requirements.txt
arrow==1.3.0
atproto>=0.0.62
Mastodon.py==1.8.1
PyTumblr==0.1.2
requests==2.32.3
tweepy>=4.15.0
python-dotenv==1.0.1
Flask==3.0.0